Remove advice before before org-capture-place template
Now that nroam inserts a non read-only \n after the sections so that text can be
inserted after nroam sections at the end of the buffer, remove the extra
read-only \n so that org-capture can insert templates.

* nroam.el (nroam--handle-org-capture): Remove function and advice.
* nroam-unlinked.el (nroam-unlinked--insert-toggle-button): Do not insert a
read-only newline as it prevents org-capture from inserting the template as it
tries to delete empty lines at the end of the buffer.